symbol / cell redraw refactor
