+++ .vscode/settings.json
@@ -1,5 +1,8 @@
 {
-    "editor.formatOnSave": true,
-    "editor.tabSize": 2,
-    "eslint.format.enable": true
+  "editor.formatOnSave": true,
+  "editor.tabSize": 2,
+  "eslint.format.enable": false,
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "eslint.lintTask.enable": true,
+  "eslint.run": "onType"
 }
