+++ board.js
@@ -17,11 +17,12 @@ export class Board {
         row.push(this.empty.copy());
         const cellDiv = document.createElement('div');
         cellDiv.classList.add('cell');
+        cellDiv.classList.add(`cell-${j}-${i}`);
         const symbolDiv = document.createElement('div');
         symbolDiv.classList.add('symbol');
         symbolDiv.innerText = '⬜';
         const counterDiv = document.createElement('div');
-        counterDiv.classList.add('symbolCounter');
+        counterDiv.classList.add('symbol-counter');
         counterDiv.innerText = '';
         cellDiv.appendChild(symbolDiv);
         cellDiv.appendChild(counterDiv);
@@ -32,14 +33,13 @@ export class Board {
     }
   }
   getSymbolDiv(x, y) {
-    return this.gridDiv.children[y].children[x].children[0];
+    return document.querySelector(`.cell-${y}-${x} .symbol`);
   }
-  updateCounter(game, x, y) {
-    const counterDiv = this.gridDiv.children[y].children[x].children[1];
-    const counter = this.cells[y][x].counter(game);
-    if (counter !== null) {
-      counterDiv.innerText = counter;
-    }
+  getCellDiv(x, y) {
+    return document.querySelector(`.cell-${y}-${x}`);
+  }
+  redrawCell(game, x, y) {
+    this.getCellDiv(x, y).replaceChildren(this.cells[y][x].render(game));
   }
   getCounterDiv(x, y) {
     return this.gridDiv.children[y].children[x].children[1];
@@ -59,52 +59,37 @@ export class Board {
   }
   async spinDiv(game, x, y, symbol) {
     await Util.delay(Util.random(600));
-    const div = this.getSymbolDiv(x, y);
-    const counterDiv = this.getCounterDiv(x, y);
-    counterDiv.innerText = '';
-    const randomSymbol = () => {
-      const set = new Set();
-      for (const symbol of Object.values(game.inventory.symbols)) {
-        set.add(symbol.emoji());
-      }
-      div.innerText = Util.randomChoose([...set]);
-    };
-    await Util.animate(div, 'startSpin', 0.1);
+    const cellDiv = this.getCellDiv(x, y);
+
+    // Rolling animation portion
+    const fakeDiv = Util.createDiv(null, 'symbol');
+    cellDiv.replaceChildren(fakeDiv);
+    await Util.animate(fakeDiv, 'startSpin', 0.1);
     for (let i = 0; i < 6; ++i) {
-      randomSymbol();
-      await Util.animate(div, 'spin', 0.12 + i * 0.02);
+      fakeDiv.innerText = game.inventory.getRandomOwnedEmoji();
+      await Util.animate(fakeDiv, 'spin', 0.12 + i * 0.02);
     }
-    div.innerText = symbol.emoji();
-    div.removeEventListener('click', div.clickEvent);
-    div.clickEvent = () => {
-      Util.drawText(game.info,
-        `${symbol.emoji()}: ${symbol.descriptionLong()}`);
-    };
-    div.addEventListener('click', div.clickEvent);
-    await Util.animate(div, 'endSpin', 0.3);
-    await Util.animate(div, 'bounce', 0.1);
-    if (symbol.counter(game) != null) {
-      counterDiv.innerText = symbol.counter(game);
-    }
+
+    // Set the actual symbol
+
+    const symbolDiv = symbol.render(game);
+    cellDiv.replaceChildren(symbolDiv);
+
+    await Util.animate(symbolDiv, 'endSpin', 0.3);
+    await Util.animate(symbolDiv, 'bounce', 0.1);
   }
   async spinDivOnce(game, x, y) {
-    const div = this.getSymbolDiv(x, y);
-    const counterDiv = this.getCounterDiv(x, y);
-    counterDiv.innerText = '';
-    await Util.animate(div, 'startSpin', 0.1);
-    const symbol = this.cells[y][x];
-    div.innerText = symbol.emoji();
-    div.removeEventListener('click', div.clickEvent);
-    div.clickEvent = () => {
-      Util.drawText(game.info,
-        `${symbol.emoji()}: ${symbol.descriptionLong()}`);
-    };
-    div.addEventListener('click', div.clickEvent);
-    await Util.animate(div, 'endSpin', 0.3);
-    await Util.animate(div, 'bounce', 0.1);
-    if (symbol.counter(game) != null) {
-      counterDiv.innerText = symbol.counter(game);
-    }
+    const cellDiv = this.getCellDiv(x, y);
+
+    // Rolling animation portion
+    const fakeDiv = Util.createDiv(null, 'symbol');
+    cellDiv.replaceChildren(fakeDiv);
+    await Util.animate(fakeDiv, 'startSpin', 0.1);
+
+    const symbolDiv = this.cells[y][x].render(game);
+
+    await Util.animate(symbolDiv, 'endSpin', 0.3);
+    await Util.animate(symbolDiv, 'bounce', 0.1);
   }
   async roll(game) {
     const symbols = [...game.inventory.symbols];
@@ -136,11 +121,11 @@ export class Board {
     await Promise.all(tasks);
   }
   async evaluate(game) {
-    this.forAllCells((cell, x, y) => {
+    this.forAllCells((cell, _, __) => {
       cell.turns++;
     });
     this.forAllCells((cell, x, y) => {
-      this.updateCounter(game, x, y);
+      this.redrawCell(game, x, y);
     });
     const evaluateRound = async (f) => {
       const tasks = [];
@@ -157,7 +142,7 @@ export class Board {
         await task();
       }
       this.forAllCells((cell, x, y) => {
-        this.updateCounter(game, x, y);
+        this.redrawCell(game, x, y);
       });
     };
     await evaluateRound((c, g, x, y) => c.evaluateConsume(g, x, y));
@@ -205,7 +190,7 @@ export class Board {
       this.cells[y][x] = sym;
       await this.spinDivOnce(game, x, y);
     }
-    this.updateCounter(game, x, y);
+    this.redrawCell(game, x, y);
   }
   async removeSymbol(game, x, y) {
     game.inventory.remove(this.cells[y][x]);
