+++ main.js
@@ -3,7 +3,8 @@ import { GameSettings } from './game_settings.js';
 import { Catalog } from './catalog.js';
 import { Board } from './board.js';
 import { Inventory } from './inventory.js';
-import { Shop } from './shop.js'; ``
+import { Shop } from './shop.js';
+``;
 import { Game } from './game.js';
 import { Progression } from './progression.js';
 
@@ -25,10 +26,16 @@ export const loadSettings = async (settings = GameSettings.instance()) => {
   return game;
 };
 
-export const loadListener = async (event) => {
+export const loadListener = async (_) => {
   document.querySelector('body').removeEventListener('click', loadListener);
   loadSettings(PROGRESSION.levelData[PROGRESSION.activeLevel]);
 };
+
+if (window.location.hash === '#dev') {
+  document.querySelectorAll('.dev-hidden').forEach((e) => {
+    e.classList.remove('dev-hidden');
+  });
+}
 
 loadSettings(PROGRESSION.levelData[PROGRESSION.activeLevel]);
 
@@ -194,7 +201,7 @@ window.simulate = async (
     const avg = (scores.reduce((acc, val) => acc + val, 0) / scores.length) | 0;
     const max = Math.max(...scores);
     const min = Math.min(...scores);
-    console.log(`${i}\tscore ${score}\tavg ${avg}\tmax ${max}`);
+    console.log(`${i}\tscore ${score}\tavg ${avg}\tmax ${max}\tmin ${min}`);
     if (game.totalTurns === 200) {
       console.log('inf!');
     }
